Refactor langchain tests
